move CICSSession to core package
