+++ packages/sdk/src/core/index.ts
@@ -9,9 +9,5 @@
  *
  */
 
-export * from "./constants";
-export * from "./core";
-export * from "./doc";
-export * from "./methods";
-export * from "./rest";
-export * from "./utils";
+
+export * from "./CICSSession";
