+++ packages/sdk/src/utils/index.ts
@@ -9,5 +9,4 @@
  *
  */
 
-export * from "./CICSSession";
 export * from "./Utils";
